Move render_html to dmscripts.html
