+++ dmscripts/html.py
@@ -1,0 +1,9 @@
+import io
+from jinja2 import Template
+
+
+def render_html(template_path, data=None):
+    with io.open(template_path, encoding='UTF-8') as htmlfile:
+        html = htmlfile.read()
+        template = Template(html)
+        return template.render(data or {})
